Move CiftiCorrelation from utils to interfaces.
