+++ xcp_d/utils/__init__.py
@@ -9,7 +9,6 @@ from xcp_d.utils.bids import (
     select_cifti_bold,
     select_registrationfile,
 )
-from xcp_d.utils.cifticonnectivity import CiftiCorrelation
 from xcp_d.utils.ciftiparcellation import CiftiParcellate
 from xcp_d.utils.ciftiresample import CiftiSurfaceResample
 from xcp_d.utils.ciftiseparatemetric import CiftiSeparateMetric
@@ -66,7 +65,6 @@ __all__ = [
     'plot_svg',
     'compute_dvars',
     'load_confound_matrix',
-    'CiftiCorrelation',
     'CiftiParcellate',
     'CiftiSeparateMetric',
     'collect_participants',
